edit:HoverableStrawberryリファクタリング
