+++ web/hooks/hooks_index.ts
@@ -182,3 +182,25 @@ export const useDelayedPreview = (delay: number) => {
 
   return showPreview;
 };
+
+export const useIsHoverable=()=>{
+  const [isHoverable, setIsHoverable] = useState(true);
+
+  useEffect(() => {
+    const hoverableQuery = window.matchMedia('(hover: hover)');
+    const updateHoverable = () => setIsHoverable(hoverableQuery.matches);
+
+    // 初回評価
+    updateHoverable();
+
+    // リスナーでリアルタイムに変化を監視
+    hoverableQuery.addEventListener('change', updateHoverable);
+
+    return () => {
+      hoverableQuery.removeEventListener('change', updateHoverable);
+    };
+  }, []);
+
+    return isHoverable;
+  }
+
